+++ packages/11ty/_includes/components/figure/audio/print.js
@@ -13,7 +13,13 @@ const path = require('path')
 module.exports = function(eleventyConfig) {
   const { imageDir } = eleventyConfig.globalData.config.figures
 
-  return function({ poster='' }) {
+  const figureCaption = eleventyConfig.getFilter('figureCaption')
+  const figureLabel = eleventyConfig.getFilter('figureLabel')
+
+  return function({ caption, credit, id, label, mediaId, mediaType, poster='' }) {
+    const labelElement = figureLabel({ caption, id, label })
+    const captionElement = figureCaption({ caption, content: labelElement, credit, mediaId, mediaType })
+
     const posterSrc = poster.startsWith('http')
       ? poster
       : path.join(imageDir, poster)
@@ -26,6 +32,7 @@ module.exports = function(eleventyConfig) {
 
     return html`
       ${imageElement}
+      ${captionElement}
     `
   }
 }