+++ packages/11ty/_includes/components/figure/video/print.js
@@ -13,10 +13,19 @@ const logger = chalkFactory('Figure Video')
  * @return     {String}  HTML containing a fallback image and a caption
  */
 module.exports = function(eleventyConfig) {
+  const figureCaption = eleventyConfig.getFilter('figureCaption')
+  const figureLabel = eleventyConfig.getFilter('figureLabel')
+
   const { imageDir } = eleventyConfig.globalData.config.figures
 
   return function({
     aspect_ratio: aspectRatio,
+    caption,
+    credit,
+    id,
+    label,
+    mediaId,
+    mediaType,
     poster=''
   }) {
     if (!poster) {
@@ -26,11 +35,14 @@ module.exports = function(eleventyConfig) {
     const posterSrc = poster.startsWith('http')
       ? poster
       : path.join(imageDir, poster)
+    const labelElement = figureLabel({ caption, id, label })
+    const captionElement = figureCaption({ caption, content: labelElement,  credit, mediaId, mediaType })
 
     return html`
       <div class="q-figure__media-wrapper--${ aspectRatio || 'widescreen' }">
         <img src="${posterSrc}" />
       </div>
+      ${captionElement}
     `
   }
 }