+++ packages/11ty/_includes/components/page-header.js
@@ -1,6 +1,8 @@
 const { html } = require('~lib/common-tags')
 const path = require('path')
 
+const checkFormat = require('../../_plugins/collections/filters/output.js')
+
 /**
  * Publication page header
  *
@@ -14,6 +16,33 @@ module.exports = function(eleventyConfig) {
   const { labelDivider } = eleventyConfig.globalData.config.pageTitle
   const { imageDir } = eleventyConfig.globalData.config.figures
 
+  const pdfConfig = eleventyConfig.globalData.config.pdf
+
+  /**
+   * @function checkPagePDF
+   * 
+   * @param {Object} config pdf object from Quire config
+   * @param {Array<string>,string,undefined} outputs outputs setting from page frontmatter 
+   * @param {bool} frontmatterSetting pdf page setting from page frontmatter
+   * 
+   * Check if the PDF link should be generated for this page
+   */
+  const checkPagePDF = (config,outputs,frontmatterSetting) => {
+
+    // Is the output being created?
+    if (!checkFormat('pdf', { data: { outputs } })) {
+      return false 
+    }
+
+    // Are the footer links set?
+    if (config.pagePDF.accessLinks.find((al) => al.header === true) === undefined)  {
+      return false
+    }
+
+    // Return the core logic check
+    return (config.pagePDF.output === true && frontmatterSetting !== false) || frontmatterSetting === true
+  }
+
   return function (params) {
     const {
       byline_format: bylineFormat,
@@ -21,7 +50,10 @@ module.exports = function(eleventyConfig) {
       label,
       pageContributors,
       subtitle,
-      title
+      title,
+      outputs,
+      page_pdf_output: pagePDFOutput,
+      key,
     } = params
 
     const classes = ['quire-page__header', 'hero']
@@ -38,7 +70,7 @@ module.exports = function(eleventyConfig) {
       ? html`
           <section
             class="${classes} hero__image"
-           style="background-image: url('${path.join(imageDir, image)}');"
+            style="background-image: url('${path.join(imageDir, image)}');"
           >
           </section>
         `
@@ -52,6 +84,18 @@ module.exports = function(eleventyConfig) {
         `
       : ''
 
+    let downloadLink = ''
+
+    if (checkPagePDF(pdfConfig,outputs,pagePDFOutput)) {
+      const text = pdfConfig.pagePDF.accessLinks.find((al) => al.header === true).label
+      const href = path.join(pdfConfig.outputDir, `${pdfConfig.filename}-${slugify(key)}.pdf`)
+      downloadLink = html`
+        <div class="quire-download" data-outputs-exclude="epub,pdf">
+          <a class="quire-download__link" href="${ href }" download><span>${ text }</span><svg class="quire-download__link__icon"><use xlink:href="#download-icon"></use></svg></a>
+        </div>
+      `
+    }
+
     return html`
       <section class="${classes}">
         <div class="hero-body">
@@ -60,6 +104,7 @@ module.exports = function(eleventyConfig) {
             ${pageTitle({ title, subtitle })}
           </h1>
           ${contributorsElement}
+          ${downloadLink}
         </div>
       </section>
       ${imageElement}