+++ packages/11ty/_plugins/transforms/outputs/pdf/transform.js
@@ -1,7 +1,10 @@
-const jsdom = require('jsdom')
+const chalkFactory = require('~lib/chalk')
 const filterOutputs = require('../filter.js')
+const jsdom = require('jsdom')
 const truncate = require('~lib/truncate')
 const writer = require('./write')
+
+const logger = chalkFactory('pdf:transform')
 
 const { JSDOM } = jsdom
 
@@ -12,9 +15,14 @@ const { JSDOM } = jsdom
  * @param      {String}  content      Output content
  * @return     {Array}   The transformed content string
  */
-module.exports = function(eleventyConfig, collections, content) {
+module.exports = async function(eleventyConfig, collections, content) {
   const pageTitle = eleventyConfig.getFilter('pageTitle')
   const slugify = eleventyConfig.getFilter('slugify')
+  const citation = eleventyConfig.getFilter('citation')
+  const citePage = eleventyConfig.getFilter('citePage')
+  const formatCitation = eleventyConfig.getFilter('formatCitation')
+  const { pdf: pdfConfig } = eleventyConfig.globalData.config
+  const slugifyIds = eleventyConfig.getFilter('slugifyIds')
 
   const writeOutput = writer(eleventyConfig)
 
@@ -34,16 +42,31 @@ module.exports = function(eleventyConfig, collections, content) {
    *
    * @param  {Object}       page     The page being transformed
    * @param  {HTMLElement}  element  HTML element on which to set data attributes
+   * @param  {boolean}      generatePagedPDF Whether to generate a PDF for this webpage
+   * 
+   * 
    */
-  const setDataAttributes = (page, element) => {
+  const setDataAttributes = (page, element, generatePagedPDF) => {
     const { dataset } = element
-    const { parentPage } = page.data
+    const { parentPage, layout } = page.data
+    const { pagePDF } = pdfConfig
 
     dataset.footerPageTitle = formatTitle(page.data)
 
     if (parentPage) {
       dataset.footerSectionTitle = formatTitle(parentPage.data)
     }
+
+    if (!generatePagedPDF) {
+      return
+    }
+
+    if (layout === 'cover') {
+      logger.warn(`${page.data.page.inputPath} uses a \`cover\` layout, this will only appear in the full publication PDF`)
+      return
+    }
+
+    dataset.pagePdf = true
   }
 
   /**
@@ -84,7 +107,114 @@ module.exports = function(eleventyConfig, collections, content) {
       item.setAttribute('href', `#${prefix}-${href.replace(/^#/, '')}`)
       item.setAttribute('id', `${prefix}-${id}`)
     })
-    // 
+  }
+
+  /**
+   * @function trimLeadingSeparator
+   * 
+   * Trims the publication URL path from @src attribs and style background-image URLs
+
+   * @param {Object} document JSDom `document` object of a section element
+   */
+  const trimLeadingSeparator = (document) => {
+    const urlPath = eleventyConfig.globalData.publication.pathname
+
+    /**
+     * This function removes either the deploy path or just the leading slash 
+     * 
+     * @example /foo/_assets/image.jpg -> _assets/image.jpg
+     * @example /_assets/image.jpg -> _assets/image.jpg
+     * @example Pass any other @src attributes (incl. `http(s)://..`)
+     * 
+     * @todo Why does background-image carry the root asset path but no pathPrefix?
+     */
+    const trimDeployPathComponentOrSlash = (srcAttr) => {
+      switch (true) {
+        case srcAttr.startsWith(urlPath):
+          return srcAttr.substr(urlPath.length)
+        case srcAttr.startsWith('/'):
+          return srcAttr.substr(1)
+        default:
+          return srcAttr
+      }
+    }
+
+    document.querySelectorAll('[src]').forEach((asset) => {
+      const src = asset.getAttribute('src')
+      asset.setAttribute('src', trimDeployPathComponentOrSlash(src))
+    })
+
+    document.querySelectorAll('[style*="background-image"]').forEach((element) => {
+      const backgroundImageUrl = element.style.backgroundImage.match(/[(](.*)[)]/)[1] || ''
+      element.style.backgroundImage = `url('${trimDeployPathComponentOrSlash(backgroundImageUrl)}')`
+    })
+  }
+
+  /**
+   * @function normalizeCoverPageData
+   * 
+   * @param {Object} pageData - page data object
+   * @param {Object} pdfConfig - configuration for the pdf
+   * 
+   * @return {Object} data formatted for the layout at _layouts/pdf-cover-pages.liquid
+   */
+  function normalizeCoverPageData(page,pdfConfig) {
+    const { pagePDFCoverPageCitationStyle } = page.data
+
+    // NB: `id` must match the @id slug scheme in `base.11ty.js` so the cover pages have the same keys
+    const accessURL = page.data.canonicalURL
+    const contributors = page.data.pageContributors ?? []    
+    const copyright = page.data.publication.copyright
+    const id = `page-${slugify(page.data.pageData.url)}` 
+
+    // @todo Need license *text* per example
+
+    const license = page.data.publication.license.name 
+
+    // @todo replace date in mla citation
+    /**
+     * The function to do this in the app client code:
+       function mlaDate(date) {
+          const options = {
+            month: 'long'
+          }
+          const monthNum = date.getMonth()
+          let month
+          if ([4, 5, 6].includes(monthNum)) {
+            let dateString = date.toLocaleDateString('en-US', options)
+            month = dateString.replace(/[^A-Za-z]+/, '')
+          } else {
+            month = (month === 8) ? 'Sept' : date.toLocaleDateString('en-US', options).slice(0, 3)
+            month += '.'
+          }
+          const day = date.getDate()
+          const year = date.getFullYear()
+          return [day, month, year].join(' ')
+        }
+     * 
+     **/
+
+    // Feed the CSL processor an access date (@todo: either make this work or use the func above..)
+    const pageCiteData = citePage({ page, context: 'page', type: 'mla' })
+    const mla = formatCitation(
+      { ...pageCiteData, accessed: '01 Oct 1999' },
+      { page, context: 'page', type: 'mla' }
+    )
+    const pageCitations = {
+      chicago: citation({ context: 'page', page, type: 'chicago' }),
+      mla
+    }
+    const title = pageTitle({ ...page.data, label: '' })
+
+    return { 
+      accessURL, 
+      citations: pageCitations, 
+      contributors, 
+      copyright, 
+      id, 
+      license, 
+      title 
+    }
   }
 
   const pdfPages = collections.pdf.map(({ outputPath }) => outputPath)
@@ -104,13 +234,16 @@ module.exports = function(eleventyConfig, collections, content) {
   const sectionElement = document.createElement('section')
   const pageId = mainElement.dataset.pageId
 
+  const hasPagePDF = (currentPage.data.page_pdf_output === true) || (pdfConfig.pagePDF.output === true && currentPage.data.page_pdf_output !== false)
+  const hasCoverPage = (currentPage.data.page_pdf_output === true) || (pdfConfig.pagePDF.output === true && currentPage.data.page_pdf_output !== false)
+
   sectionElement.innerHTML = mainElement.innerHTML
 
   for (const className of mainElement.classList) {
     sectionElement.classList.add(className)
   }
 
-  setDataAttributes(currentPage, sectionElement)
+  setDataAttributes(currentPage, sectionElement, hasPagePDF)
 
   // set an id for anchor links to each section
   sectionElement.setAttribute('id', pageId)
@@ -121,10 +254,17 @@ module.exports = function(eleventyConfig, collections, content) {
   // prefix footnote attributes to prevent duplicates
   prefixFootnotes(sectionElement, pageId)
 
-  // remove non-pdf content
-  filterOutputs(sectionElement, 'pdf')
-  collections.pdf[pageIndex].svgSymbolElements = Array.from(svgSymbolElements)
-  collections.pdf[pageIndex].sectionElement = sectionElement
+  // Final cleanups: remove non-pdf content, remove image leading slashes, slugify it all
+  filterOutputs(sectionElement, 'pdf')  
+  trimLeadingSeparator(sectionElement)
+  slugifyIds(sectionElement)
+
+  collections.pdf[pageIndex].svgSymbolElements = Array.from(svgSymbolElements).map( el => el.outerHTML )
+  collections.pdf[pageIndex].sectionElement = sectionElement.outerHTML
+
+  if (hasPagePDF && hasCoverPage) {
+    collections.pdf[pageIndex].coverPageData = normalizeCoverPageData(currentPage,pdfConfig)         
+  }
 
   /**
    * Once this transform has been called for each PDF page